\name{split_example}
\alias{split_example}
\docType{data}
\title{Simulated data for split-questionnaire examples}
\description{A simulated complete data set used to demonstrate pattern assignment and planned missingness.}
\usage{split_example}
\format{A data frame with 300 observations and five numeric variables.}
\source{Generated by \file{data-raw/create_example.R}.}
